+++ v2/src/modules/txs/ModuleTxDetails.vue
@@ -12,14 +12,14 @@
                             <div v-else>
                                 <app-chip :bg="titleBg" rounded="xl" text="" class="">
                                     <p>
-                                        {{ titleStatus }}
+                                        {{ titleStatusString }}
                                         <span>
                                             <v-progress-circular
                                                 v-if="txStatus === TxStatus.pending"
                                                 indeterminate
                                                 color="white"
                                                 size="15"
-                                                width="3"
+                                                width="2"
                                                 class="ml-2"
                                             />
                                         </span>
@@ -307,6 +307,18 @@ const titleStatus = computed<TitleStatus>(() => {
     }
     return TitleStatus.replaced
 })
+const titleStatusString = computed<string>(() => {
+    switch (titleStatus.value) {
+        case TitleStatus.success:
+            return t('txs.details.status.successful')
+        case TitleStatus.failed:
+            return t('txs.details.status.failed')
+        case TitleStatus.pending:
+            return t('txs.details.status.pending')
+        default:
+            return t('txs.details.status.replaced')
+    }
+})
 
 const titleBg = computed<string>(() => {
     switch (titleStatus.value) {
@@ -314,8 +326,10 @@ const titleBg = computed<string>(() => {
             return 'success'
         case TitleStatus.failed:
             return 'error'
+        case TitleStatus.pending:
+            return 'purple'
         default:
-            return 'warning'
+            return 'info'
     }
 })
 
@@ -323,7 +337,7 @@ const timestamp = computed<string>(() => {
     if (transactionData.value && transactionData.value?.timestamp) {
         const date = new Date(transactionData.value.timestamp * 1e3).toLocaleDateString()
         const time = new Date(transactionData.value.timestamp * 1e3).toTimeString().split('GMT')[0]
-        const timeago = timeAgo(new Date(transactionData.value.timestamp * 1e3))
+        const timeago = timeAgo(new Date(transactionData.value.timestamp * 1e3), false)
         const [month, day, year] = date.split('/')
         return `${year}-${month}-${day}, ${time}, ${timeago}`
     }