+++ v2/src/views/ViewPortfolio.vue
@@ -31,7 +31,7 @@
                             <address-balance-totals
                                 :title="$t('block.tokenBalance')"
                                 :is-loading="!store.portfolioTokensIsLoaded()"
-                                :subtext="`${tokensCount} ${$t('block.tokenTotal')}`"
+                                :subtext="tokenCountSubtext"
                                 :balance="tokenBalanceFiat.value"
                                 :balance-tooltip="tokenBalanceFiat.tooltipText"
                             />
@@ -46,7 +46,7 @@
                 <address-balance-totals
                     :title="$t('block.tokenBalance')"
                     :is-loading="!store.portfolioTokensIsLoaded()"
-                    :subtext="`${tokensCount} total tokens`"
+                    :subtext="tokenCountSubtext"
                     :balance="tokenBalanceFiat.value"
                     :balance-tooltip="tokenBalanceFiat.tooltipText"
                 />
@@ -91,7 +91,7 @@ import { reactive, computed } from 'vue'
 import { useStore } from '@/store'
 import { useDisplay } from 'vuetify/lib/framework.mjs'
 import { onBeforeRouteUpdate } from 'vue-router'
-import { formatUsdValue, formatNumber, FormattedNumber } from '@core/helper/number-format-helper'
+import { formatUsdValue, FormattedNumber } from '@core/helper/number-format-helper'
 import { TokenSort, TOKEN_FILTER_VALUES } from '@module/address/models/TokenSort'
 import { MarketDataFragment } from '@core/composables/CoinData/getLatestPrices.generated'
 import { useCoinData } from '@core/composables/CoinData/coinData.composable'
@@ -152,10 +152,6 @@ const tokenBalanceFiat = computed<FormattedNumber>(() => {
     return formatUsdValue(store.portfolioTokensFiatBN)
 })
 
-const tokensCount = computed<string>(() => {
-    return formatNumber(store.portfolioTokensRaw().length)
-})
-
 const { getEthereumTokensMap } = useCoinData()
 
 const tokenIcons = computed<string[]>(() => {
@@ -183,6 +179,10 @@ const tokenIcons = computed<string[]>(() => {
             .filter(i => i !== '')
     }
     return []
+})
+
+const tokenCountSubtext = computed<string>(() => {
+    return store.portfolioTokensRaw().length === 1 ? t('token.tokenTotal') : t('token.tokenTotal', { n: store.portfolioTokensRaw().length })
 })
 </script>
 <style lang="scss">